some modification

+++ WebRoot/WEB-INF/jsp/economy/list.jsp
@@ -83,7 +83,7 @@
                                     <s:date name="date" format="yyyy-MM-dd"/>
                                 </td>
 					            <td><s:a type="button" cssClass="btn btn-xs btn-primary" href="economy_editUI?id=%{id}">修改</s:a></td>
-	                 			 <td><s:a type="button" cssClass="btn btn-xs btn-primary" href="economy_delete?id=%{id}">删除</s:a></td>
+	                 			 <td><s:a type="button" cssClass="btn btn-xs btn-primary" onclick="return confirm('你确定要删除吗')" href="economy_delete?id=%{id}">删除</s:a></td>
                             </tr>
                         </s:iterator>
